pyproject.toml
[project]
name = "dia1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.3",
    "streamlit>=1.39.0",
]
